Address comments on PR
