+++ python/ray/llm/_internal/serve/configs/constants.py
@@ -17,3 +17,7 @@ if DEFAULT_MULTIPLEX_DOWNLOAD_TIMEOUT_S <= 0:
 DEFAULT_MULTIPLEX_DOWNLOAD_TRIES = int(
     os.getenv("DEFAULT_MULTIPLEX_DOWNLOAD_RETRIES", "3")
 )
+
+DEFAULT_TARGET_ONGOING_REQUESTS = 16
+
+FALLBACK_MAX_ONGOING_REQUESTS = 64
