+++ python/ray/llm/tests/batch/gpu/stages/test_vllm_engine_stage.py
@@ -14,38 +14,6 @@ from ray.llm._internal.batch.stages.vllm_engine_stage import (
 from ray.util.scheduling_strategies import PlacementGroupSchedulingStrategy
 
 
-@pytest.fixture(scope="module")
-def download_model_ckpt():
-    """
-    Download the model checkpoint and tokenizer from S3 for testing
-    The reason to download the model from S3 is to avoid downloading the model
-    from HuggingFace hub during testing, which is flaky because of the rate
-    limit and HF hub downtime.
-    """
-
-    REMOTE_URL = "https://air-example-data.s3.amazonaws.com/facebook-opt-125m/"
-    FILE_LIST = [
-        "config.json",
-        "flax_model.msgpack",
-        "generation_config.json",
-        "merges.txt",
-        "pytorch_model.bin",
-        "special_tokens_map.json",
-        "tokenizer_config.json",
-        "vocab.json",
-    ]
-
-    # Create a temporary directory and save the model and tokenizer.
-    with tempfile.TemporaryDirectory() as checkpoint_dir:
-        # Download the model checkpoint and tokenizer from S3.
-        for file_name in FILE_LIST:
-            response = requests.get(REMOTE_URL + file_name)
-            with open(os.path.join(checkpoint_dir, file_name), "wb") as fp:
-                fp.write(response.content)
-
-        yield os.path.abspath(checkpoint_dir)
-
-
 @pytest.fixture
 def mock_vllm_wrapper():
     with patch(
