+++ python/ray/llm/tests/serve/configs/configs/matching_configs/hf_prompt_format.yaml
@@ -6,11 +6,11 @@ model_loading_config:
   model_id: mistral-community/pixtral-12b
   model_source: "/home/ray/tests/rayllm/backend/server/configs/cached_model_processors/mistral-community--pixtral-12b"
 
-generation_config:
-  prompt_format:
-    use_hugging_face_chat_template: true
-  stopping_sequences: []
-  stopping_tokens: [128001, 128009]  # <|end_of_text|>, <|eot_id|>
+# generation_config:
+#   prompt_format:
+#     use_hugging_face_chat_template: true
+#   stopping_sequences: []
+#   stopping_tokens: [128001, 128009]  # <|end_of_text|>, <|eot_id|>
 
 llm_engine: VLLMEngine
 
@@ -36,3 +36,5 @@ deployment_config:
   autoscaling_config:
     target_ongoing_requests: 32
   max_ongoing_requests: 1
+
+
