--- python/ray/llm/tests/serve/cpu/common/test_hf_prompt.py
@@ -1,112 +1,0 @@
-import pytest
-
-from ray.llm._internal.serve.configs.prompt_formats import (
-    ContentContent,
-    HuggingFacePromptFormat,
-    ImageContent,
-    Message,
-    Prompt,
-)
-
-
-@pytest.fixture
-def prompt_format(download_model_ckpt):
-    hf_prompt_format = HuggingFacePromptFormat(use_hugging_face_chat_template=True)
-    hf_prompt_format.set_processor(model_id_or_path=download_model_ckpt)
-    return hf_prompt_format
-
-
-def test_prompt_format_on_string_message(prompt_format):
-    messages = Prompt(prompt="This is a test message.")
-    with pytest.raises(ValueError):
-        prompt_format.generate_prompt(messages=messages)
-
-
-def test_prompt_format_on_prompt_object(prompt_format):
-    messages = Prompt(
-        prompt=[
-            Message(role="system", content="You are a helpful assistant."),
-            Message(
-                role="user",
-                content=[
-                    ContentContent(field="text", content="Can this animal"),
-                    ImageContent(
-                        field="image_url",
-                        image_url={"url": "https://example.com/dog.jpg"},
-                    ),
-                    ContentContent(field="text", content="live here?"),
-                    ImageContent(
-                        field="image_url",
-                        image_url={"url": "https://example.com/mountain.jpg"},
-                    ),
-                ],
-            ),
-            Message(
-                role="assistant",
-                content="It looks like you've shared an image of a "
-                "dog lying on a wooden floor, and another "
-                "image depicting a serene landscape with a "
-                "sunset over a snowy hill or mountain.",
-            ),
-            Message(
-                role="user",
-                content="So you are suggesting you can find a poppy living in the snowy mountain?",
-            ),
-        ],
-    )
-
-    formated_prompt = prompt_format.generate_prompt(messages=messages)
-    assert formated_prompt.text == (
-        "<s>[INST]Can this animal[IMG]live here?[IMG][/INST]It looks like you've "
-        "shared an image of a dog lying on a wooden floor, and another image "
-        "depicting a serene landscape with a sunset over a snowy hill or "
-        "mountain.</s>[INST]You are a helpful assistant.\n\nSo you are suggesting "
-        "you can find a poppy living in the snowy mountain?[/INST]"
-    )
-    assert len(formated_prompt.image) == 2
-    assert formated_prompt.image[0].image_url == "https://example.com/dog.jpg"
-    assert formated_prompt.image[1].image_url == "https://example.com/mountain.jpg"
-
-
-def test_prompt_format_on_list_of_messages(prompt_format):
-    messages = [
-        Message(role="system", content="You are a helpful assistant."),
-        Message(
-            role="user",
-            content=[
-                ContentContent(field="text", content="Can this animal"),
-                ImageContent(
-                    field="image_url",
-                    image_url={"url": "https://example.com/dog.jpg"},
-                ),
-                ContentContent(field="text", content="live here?"),
-                ImageContent(
-                    field="image_url",
-                    image_url={"url": "https://example.com/mountain.jpg"},
-                ),
-            ],
-        ),
-        Message(
-            role="assistant",
-            content="It looks like you've shared an image of a "
-            "dog lying on a wooden floor, and another "
-            "image depicting a serene landscape with a "
-            "sunset over a snowy hill or mountain.",
-        ),
-        Message(
-            role="user",
-            content="So you are suggesting you can find a poppy living in the snowy mountain?",
-        ),
-    ]
-
-    formated_prompt = prompt_format.generate_prompt(messages=messages)
-    assert len(formated_prompt.image) == 2
-    assert formated_prompt.text == (
-        "<s>[INST]Can this animal[IMG]live here?[IMG][/INST]It looks like you've "
-        "shared an image of a dog lying on a wooden floor, and another image "
-        "depicting a serene landscape with a sunset over a snowy hill or "
-        "mountain.</s>[INST]You are a helpful assistant.\n\nSo you are suggesting "
-        "you can find a poppy living in the snowy mountain?[/INST]"
-    )
-    assert formated_prompt.image[0].image_url == "https://example.com/dog.jpg"
-    assert formated_prompt.image[1].image_url == "https://example.com/mountain.jpg"
